Split voronoi generator in onw module, isort, black
